BigQuery: Dry run big query.
Big query has an important feature of dry running the query before
submitting it this is important as you pay par query, and a bad
query can cost quite a bit of money.
Besides that dry run also does validation to tell you if the query
is going to fail, and why.