BigQuery: Add preview button.
Previewing data on BigQuery should only be done using the dedicated API
this is because running a select * query, even with a limit can still
scan a lot of data, and cost a lot of money.

The preview button is added to each table in the schema browser where
supported(currently only in BigQuery).